feat(spawn): Attach spawn animation to a tank
Closes #32
