+++ src/core/DisplayObject.js
@@ -28,6 +28,12 @@ class DisplayObject {
     return new BoundingBox(min, max);
   }
 
+  /**
+   * Called on each game loop iteration
+   */
+  // eslint-disable-next-line class-methods-use-this
+  update() {}
+
   // Must-have for each render object
   render() {
     return {
